# openapi.yml
openapi: 3.0.3
info:
  title: PrimeraApiCasa - Biblioteca API
  description: |
    REST API para gestión de una biblioteca. Usa autenticación JWT.

    **Roles:**
    - `USER` — puede consultar el catálogo (autores, libros, detalles, categorías, libro-categorías)
    - `ADMIN` — acceso completo, incluyendo escritura en el catálogo y gestión de roles y usuarios

    **Autenticación:** Los endpoints protegidos requieren el header `Authorization: Bearer <token>`.
    El token se obtiene en `/auth/login` o `/auth/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080/micro
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /roles:
    get:
      tags: [Roles]
      summary: Listar todos los roles
      description: Requiere rol `ADMIN`.
      responses:
        '200':
          description: Lista de roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Roles]
      summary: Crear un rol
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolCreateRequest'
            example:
              nombre: MODERATOR
      responses:
        '201':
          description: Rol creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /roles/{id}:
    get:
      tags: [Roles]
      summary: Obtener rol por ID
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Rol encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Roles]
      summary: Actualizar un rol
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolCreateRequest'
            example:
              nombre: MODERATOR
      responses:
        '200':
          description: Rol actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Roles]
      summary: Eliminar un rol
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Rol eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── USUARIO-ROLES ────────────────────────────────────────────────────────────
  /usuario-roles/usuario/{usuarioId}:
    get:
      tags: [UsuarioRoles]
      summary: Obtener roles de un usuario
      description: Requiere rol `ADMIN`.
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del usuario
      responses:
        '200':
          description: Lista de asignaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuario-roles/rol/{rolId}:
    get:
      tags: [UsuarioRoles]
      summary: Obtener usuarios con un rol
      description: Requiere rol `ADMIN`.
      parameters:
        - name: rolId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del rol
      responses:
        '200':
          description: Lista de asignaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuario-roles:
    post:
      tags: [UsuarioRoles]
      summary: Asignar rol a usuario
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRolCreateRequest'
            example:
              usuarioId: 1
              rolId: 2
      responses:
        '201':
          description: Rol asignado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioRolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuario-roles/{usuarioId}/{rolId}:
    delete:
      tags: [UsuarioRoles]
      summary: Remover rol de usuario
      description: Requiere rol `ADMIN`.
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del usuario
        - name: rolId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del rol
      responses:
        '204':
          description: Rol removido
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

# ─── COMPONENTS ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: ID del recurso
  responses:
    Unauthorized:
      description: Token ausente o inválido
    NotFound:
      description: Recurso no encontrado
    Forbidden:
      description: No tienes permisos para este recurso
  schemas:
...
    RolCreateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          example: ADMIN
    RolResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: ADMIN

    # UsuarioRoles
    UsuarioRolCreateRequest:
      type: object
      required: [usuarioId, rolId]
      properties:
        usuarioId:
          type: integer
          format: int64
          example: 1
        rolId:
          type: integer
          format: int64
          example: 2
    UsuarioRolResponse:
      type: object
      properties:
        usuarioId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: juan123
        rolId:
          type: integer
          format: int64
          example: 2
        rolNombre:
          type: string
          example: ADMIN