# src/main/java/com/breaze/PrimeraApiCasa/dto/RolCreateRequest.java
package com.breaze.PrimeraApiCasa.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class RolCreateRequest {

    private String nombre;
}