# src/main/java/com/breaze/PrimeraApiCasa/dto/RolResponse.java
package com.breaze.PrimeraApiCasa.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class RolResponse {

    private Long id;
    private String nombre;
}