# src/main/java/com/breaze/PrimeraApiCasa/dto/UsuarioRolCreateRequest.java
package com.breaze.PrimeraApiCasa.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class UsuarioRolCreateRequest {

    private Long usuarioId;
    private Long rolId;
}